Add files via upload
This list contains the top 20,000 spoken English words in order, the list contains certain errors, some words are not included due to profanity or unrecognised symbols (for instance the word "don't" will be listed as "don" and "t". There is also words like OYMYO that is used in the markup but are not measured in the spoken words.